+++ src/commands/mod.rs
@@ -1,3 +1,4 @@
+mod listen;
 mod nodes;
 mod send;
 
@@ -27,6 +28,7 @@ pub trait Command {
 pub fn create_command(command: &Commands) -> Result<Box<dyn Command>, CliError> {
     match command {
         Commands::Nodes => Ok(Box::new(nodes::NodesCommand)),
+        Commands::Listen => Ok(Box::new(listen::ListenCommand)),
         Commands::Send {
             message,
             dest,